Fix Windows nighty builds. (#4300)
Fixes #4299, build failure on Windows due to warnings treated as errors:
```
D:\a\TileDB\TileDB\tiledb\sm\subarray\test\unit_range_subset.cc(376,32): error C2220: the following warning is treated as an error [D:\a\TileDB\TileDB\build\tiledb\tiledb\sm\subarray\test\unit_range_subset.vcxproj] 
```
---
TYPE: NO_HISTORY

+++ tiledb/api/c_api/vfs/test/unit_capi_vfs.cc
@@ -591,11 +591,13 @@ TEST_CASE("C API: tiledb_vfs_read argument validation", "[capi][vfs]") {
     CHECK(tiledb_status(rc) == TILEDB_ERR);
   }
   SECTION("invalid offset") {
-    auto rc{tiledb_vfs_read(x.ctx, x.vfs_fh, -1, buffer, 0)};
+    auto rc{
+        tiledb_vfs_read(x.ctx, x.vfs_fh, static_cast<uint64_t>(-1), buffer, 0)};
     CHECK(tiledb_status(rc) == TILEDB_ERR);
   }
   SECTION("invalid number of bytes") {
-    auto rc{tiledb_vfs_read(x.ctx, x.vfs_fh, 0, buffer, -1)};
+    auto rc{
+        tiledb_vfs_read(x.ctx, x.vfs_fh, 0, buffer, static_cast<uint64_t>(-1))};
     CHECK(tiledb_status(rc) == TILEDB_ERR);
   }
 }
